+++ src/components/screens/Opportunity/index.jsx
@@ -8,9 +8,11 @@ import Loader from 'react-loader-spinner';
 import classNames from 'classnames';
 import { compose } from 'redux';
 import {
+  UpdateNewBid,
   expandAllSectionsAction,
   getOpportunity,
-  onGetValidatedProposalDetails
+  onGetValidatedProposalDetails,
+  closeNewbidflags
 } from '../../../redux/actions/proposal-actions';
 import { onRefreshUserData } from '../../../redux/actions/sso-auth-actions';
 import {
@@ -18,7 +20,8 @@ import {
   getPendingValidatedItems,
   getProposalDetails,
   getSelectedBid,
-  isProposalLoading
+  isProposalLoading,
+  getStatusOfNewBid
 } from '../../../redux/selectors';
 import Toolbar from '../../views/toolbar';
 import MatomoHOC from '../../HOC/MatomoHOC';
@@ -26,6 +29,10 @@ import UnityFooter from '../../common/Footer';
 import UnityGrid from '../../common/atoms/inputs/Grid';
 import UnityTab from '../../common/atoms/inputs/Tab';
 import { onHandleOpenClose } from '../../../redux/actions/sidebar-actions';
+import { SOCKET_URL } from '../../../constants/api'
+import Modal from 'apollo-react/components/Modal';
+import ProcessingCRM from '../../views/modals/ProcessingCRM';
+import BidDoneBanner from '../../views/BidDoneBanner';
 
 type State = {
   selectedView: string
@@ -45,19 +52,47 @@ type Props = {
   trackEvent: any,
   trackPageView: any,
   proposalDetail: any,
-  getOpportunityInfo: (oppId: string) => void
+  getOpportunityInfo: (oppId: string, flag?: boolean) => void
 };
 
 export class Opportunity extends Component<Props, State> {
   toRef;
-
   constructor(props: Object) {
     super(props);
-
     this.state = {
       selectedView: 'questions',
-      enableValidateTab: false
+      enableValidateTab: false,
+      newbidstatus: false
     };
+  }
+  connectsocket(){
+    const {
+      match: { params },
+      AddNewBid,
+      getOpportunityInfo
+    } = this.props;
+    this.socketconnection = null;
+    this.socketconnection = new WebSocket(SOCKET_URL);
+    this.socketconnection.onopen =  (event) => {
+      console.log('socket connected',event)
+      if(params.id){
+        this.socketconnection.send(JSON.stringify({
+          action: 'ADD_OPPORTUNITY',
+          body: {oppId : params.id}
+        }));
+      }
+    };
+
+    this.socketconnection.addEventListener('message',  async (response) =>{
+      let data = JSON.parse(response.data);
+      console.log('data.event :>> ', data.event);
+       if(data.event == 'IN_PROGRESS'){
+         AddNewBid(data.data);
+       }
+       if(data.event == 'COMPLETED'){
+        getOpportunityInfo(params.id, true);
+       }
+    });
   }
 
   componentDidMount() {
@@ -69,10 +104,11 @@ export class Opportunity extends Component<Props, State> {
       expandAllSections,
       trackPageView,
       eventCategories,
+      AddNewBid,
       location: { search },
       match: { params }
     } = this.props;
-
+    this.connectsocket();
     expandAllSections(false);
     let selectedView = new URLSearchParams(search).get('viewType');
     if (selectedView && selectedView == "documents") this.setState({ selectedView });
@@ -100,6 +136,15 @@ export class Opportunity extends Component<Props, State> {
   }
 
   componentWillUnmount() {
+    const { handleOpenClose} = this.props;
+    this.socketconnection.send(JSON.stringify({
+      action: '$disconnect',
+      body: {}
+    }));
+    this.socketconnection.close();
+    if(handleOpenClose)
+     handleOpenClose(false);
+
     localStorage.removeItem('proposalTypeView');
     localStorage.removeItem('proposalId');
 
@@ -158,9 +203,10 @@ export class Opportunity extends Component<Props, State> {
       isLoading,
       details,
       isOpen,
+      selectedBid,
       match: { params }
     } = this.props;
-
+    const { bidStatus } = selectedBid.toJS();
     if (isLoading)
       return (
         <div className="proposal-loader">
@@ -170,21 +216,15 @@ export class Opportunity extends Component<Props, State> {
 
     return (
       <div className="proposal-details">
-        <UnityGrid data={details} isOpen={isOpen} />
+        <UnityGrid data={details} isOpen={isOpen} bidStatus={bidStatus} />
         <UnityTab id={params.id} enableValidateTab={enableValidateTab} selectedView={selectedView}/>
       </div>
     );
   };
 
-  componentWillUnmount(){
-    const { handleOpenClose} = this.props;
-    if(handleOpenClose)
-     handleOpenClose(false);
-  }
-
   render() {
-    const { isSidebarOpen, selectedBid } = this.props;
-    const { questionTemplateVersionNumber, opportunityType  } = selectedBid.toJS();
+    const { isSidebarOpen, selectedBid, newbidflag, closeNewbidflag } = this.props;
+    const { questionTemplateVersionNumber, opportunityType, bidStatus } = selectedBid.toJS();
     return (
       <div
         className={classNames('proposal-wrapper', {
@@ -192,7 +232,14 @@ export class Opportunity extends Component<Props, State> {
         })}
       >
         <Toolbar />
+        <BidDoneBanner
+          isOpen={newbidflag}
+          onCloseHandler = {()=> closeNewbidflag()}
+        />
         {this.renderContent()}
+        {
+          bidStatus && <ProcessingCRM isOpen={bidStatus} />
+        }
         <UnityFooter
           questionTemplateVersionNumber={questionTemplateVersionNumber || ''}
           opportunityType={opportunityType || ''}
@@ -209,7 +256,8 @@ const mapStateToProps = (state: Map) => ({
   notifications: getPendingValidatedItems(state),
   proposalDetail: getProposalDetails(state),
   isOpen: getIsOpen(state),
-  selectedBid: getSelectedBid(state)
+  selectedBid: getSelectedBid(state),
+  newbidflag: getStatusOfNewBid(state)
 });
 
 export default compose(
@@ -220,5 +268,7 @@ export default compose(
     getValidatedData: onGetValidatedProposalDetails,
     handleOpenClose: onHandleOpenClose,
     expandAllSections: expandAllSectionsAction,
+    AddNewBid: UpdateNewBid,
+    closeNewbidflag: closeNewbidflags
   })
 )(MatomoHOC(Opportunity));